Feat: readme 수정 기능 컨트롤러 생성

See merge request s10-final/S10P31E101!51

+++ backend/src/main/java/com/dev101/coa/domain/repo/entity/Comment.java
@@ -21,10 +21,10 @@ public class Comment extends BaseEntity {
     private RepoView repoView;
 
     @Column(name = "comment_start_index")
-    private Short commentStartIndex;
+    private Integer commentStartIndex;
 
     @Column(name = "comment_end_index")
-    private Short commentEndIndex;
+    private Integer commentEndIndex;
 
     @Column(name = "comment_content", length = 255)
     private String commentContent;
